Data.Trees.KdTree

Documentation

data Point3d

Constructors

Point3d

Fields

p3x :: Double

p3y :: Double

p3z :: Double

data KdTree point

Constructors

KdNode

Fields

kdLeft :: KdTree point

kdPoint :: point

kdRight :: KdTree point

kdAxis :: Int

KdEmpty

fromList :: Point p e => [p] -> KdTree p

fromListWithDepth :: Point p e => [p] -> Int -> KdTree p

fromListWithDepth selects an axis based on depth so that the axis cycles through all valid values.

addPoint :: Point p e => p -> KdTree p -> KdTree p

addPoint inserts a point into a tree

rebalance :: Point p e => KdTree p -> KdTree p

rebalance a tree

toList :: KdTree p -> [p]

subtrees :: KdTree p -> [KdTree p]

subtrees t returns a list containing t and all its subtrees, including the empty leaf nodes.

nearestNeighbor :: Point p e => KdTree p -> p -> Maybe p

nearestNeighbor tree p returns the nearest neighbor of p in tree.

nearNeighbors :: Point p e => KdTree p -> e -> p -> [p]

nearNeighbors tree p returns all neighbors within distance r from p in tree.

isValid :: Point p e => KdTree p -> Bool

isValid tells whether the K-D tree property holds for a given tree. Specifically, it tests that all points in the left subtree lie to the left of the plane, p is on the plane, and all points in the right subtree lie to the right.

allSubtreesAreValid :: Point p e => KdTree p -> Bool

allSubtreesAreValid tells whether the K-D tree property holds for the given tree and all subtrees.

kNearestNeighbors :: (Eq p, Point p e) => KdTree p -> Int -> p -> [p]

kNearestNeighbors tree k p returns the k closest points to p within tree.

remove :: (Eq p, Point p e) => KdTree p -> p -> KdTree p

remove t p removes the point p from t.